Reorganise dataset.py and break some functions out into utils.py
